pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikita Aspaev",email = "asspaev@yandex.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn[standart] (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "orjson (>=3.10.15,<4.0.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "ipywidgets (>=8.1.6,<9.0.0)",
    "curl-cffi (>=0.10.0,<0.11.0)",
    "joblib (>=1.4.2,<2.0.0)",
    "lightgbm (>=4.6.0,<5.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "lxml (>=5.4.0,<6.0.0)",
    "cloudpickle (>=3.1.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
